keyboard.cpp and settings.h: fix small bug (using signed char instead of int)

+++ gui/source/settings.h
@@ -94,9 +94,9 @@ typedef struct _Settings_t {
 		s8 extvram;
 		s8 lockarm9scfgext;
 		s8 donor;
-		s8 red;
-		s8 green;
-		s8 blue;
+		int red;
+		int green;
+		int blue;
 	} pergame;
 } Settings_t;
 extern Settings_t settings;
